requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
websockets>=11.0.3
pydantic>=2.0.0
python-dotenv==0.19.0
peewee>=3.16.0
psycopg2-binary>=2.9.0
python-logging>=0.4.9
PyJWT>=2.6.0
google-auth>=2.16.0
